refact : LineAcceptanceTest 중복코드 제거 및 Request 분리
