+++ src/test/java/nextstep/subway/request/MethodType.java
@@ -13,12 +13,18 @@ public enum MethodType {
         }
         return R.post(W.getUri());
     }),
-    DELETE(((R, W) -> {
+    DELETE((R, W) -> {
         if (W.getPathParams() != null) {
             return R.delete(W.getUri(), W.getPathParams());
         }
         return R.delete(W.getUri());
-    }));
+    }),
+    PUT((R, W) -> {
+        if (W.getPathParams() != null) {
+            return R.put(W.getUri(), W.getPathParams());
+        }
+        return R.put(W.getUri());
+    });
 
     WhenFunction function;
